image_analyzer.py
import cv2
import numpy as np
from PIL import Image
from typing import Dict
import tensorflow as tf
from logger import logger


class ImageAnalyzer:
    def __init__(self):
        # Load any models for image-based candle recognition if needed
        # For now, no external model used; stub implementation
        pass

    def preprocess_image(self, image_path: str) -> np.ndarray:
        """
        Load image, convert to grayscale, resize to model input size if needed.
        """
        try:
            image = Image.open(image_path).convert('RGB')
            image_np = np.array(image)
            # Convert to grayscale
            gray = cv2.cvtColor(image_np, cv2.COLOR_RGB2GRAY)
            # Resize to fixed size e.g. 224x224 for model input - placeholder
            resized = cv2.resize(gray, (224, 224))
            # Normalize
            normalized = resized.astype(np.float32) / 255.0
            # Expand dims for model input
            input_tensor = np.expand_dims(normalized, axis=(0, -1))  # Shape (1,224,224,1)
            return input_tensor
        except Exception as e:
            logger.error(f"Error preprocessing image {image_path}: {e}")
            raise

    def analyze_chart_image(self, image_path: str) -> Dict[str, float]:
        """
        Analyze screenshot chart image and extract signals.
        Returns dict with signals and confidence.
        """
        try:
            input_tensor = self.preprocess_image(image_path)
            # Placeholder for actual model prediction
            # For demo, randomly assign signal
            import random
            signals = ["BUY", "SELL", "HOLD"]
            signal = random.choice(signals)
            confidence = round(random.uniform(0.7, 0.95), 2)

            return {
                "signal": signal,
                "confidence": confidence,
                "source": "image_analyzer"
            }
        except Exception as e:
            logger.error(f"Error analyzing chart image: {e}")
            return {
                "signal": "HOLD",
                "confidence": 0.0,
                "source": "image_analyzer"
            }


image_analyzer = ImageAnalyzer()
