requirements.txt
python-telegram-bot==20.3
tensorflow==2.12.0
opencv-python-headless==4.7.0.72
Pillow==9.5.0
pandas==2.0.2
requests==2.31.0
selenium==4.9.1
python-dotenv==1.0.0
numpy==1.24.3
